worker/requirements.txt
openai-whisper
moviepy==1.0.3
torch
decorator
langchain_community
requests